Clippy cleanup: iterate map values, drop redundant cast, move helper fn
Minor cleanups turned up by `cargo clippy --tests` on the three
distributed crates. None affects behavior.

- shard_manager.rs: three `for_kv_map` violations in assign_shards
  tests — replaced `for (_, v) in &map` with `map.values()`.
- merge.rs: the unnecessary `as i32` cast in
  test_merge_performance_regression (range iterator already yields
  i32).
- worker/service.rs: `local_request_to_descriptor` was defined after
  the `#[cfg(test)] mod tests` block, which triggers the
  `items-after-test-module` lint. Moved it above the test module so
  the source order matches the conventional "impls → helpers → tests"
  layout.

Result: the three distributed crates are now clippy-clean in
both `--lib` and `--lib --tests` modes.

+++ crates/coordinator/src/merge.rs
@@ -262,7 +262,7 @@ mod tests {
         let n = 1000;
         let num_batches = 10;
         let batches: Vec<RecordBatch> = (0..num_batches).map(|b| {
-            let ids: Vec<i32> = (0..n).map(|i| (b * n + i) as i32).collect();
+            let ids: Vec<i32> = (0..n).map(|i| b * n + i).collect();
             let distances: Vec<f32> = (0..n).map(|i| (i as f32) * 0.001 + (b as f32) * 0.01).collect();
             make_distance_batch(ids, distances)
         }).collect();
